refactor: add DaysOfWeek component
