+++ src/components/DatePicker/components/MonthSelector.tsx
@@ -2,19 +2,19 @@ import { FC } from 'react'
 import { Option } from '../DatePicker.types'
 import { Dropdown } from './Dropdown'
 import { getMonthsByLocale } from '../../../utils/dates'
+import { useDatePicker } from '../use-datepicker'
 
 interface MonthSelectorProps {
-  language: string
   selectedMonth: number
   onChange: (value: string) => void
 }
 
 export const MonthSelector: FC<MonthSelectorProps> = ({
-  language = 'en',
   selectedMonth,
   onChange,
 }) => {
-  const months = getMonthsByLocale(language)
+  const { state } = useDatePicker()
+  const months = getMonthsByLocale(state.language)
 
   const monthOptions: Option[] = months.map((month, index) => ({
     label: month,
